+++ src/FunFair.BuildCheck.Runner/Setup.cs
@@ -100,7 +100,10 @@ internal static class Setup
             .AddProjectCheck<EnableMicrosoftExtensionsConfigurationBinderSourceGeneratorPolicy>()
             .AddProjectCheck<JsonSerializerIsReflectionEnabledByDefaultPolicy>()
             .AddProjectCheck<OptimizationPreferencePolicy>()
-            .AddProjectCheck<GenerateSbomPolicy>();
+            .AddProjectCheck<GenerateSbomPolicy>()
+            .AddProjectCheck<TestHarnessExeProjectsMustSetIsTestProjectToFalse>()
+            .AddProjectCheck<TestHarnessExeProjectsMustSetIsTestingPlatformApplicationToFalse>()
+            .AddProjectCheck<TestHarnessExeProjectsMustSetTestingPlatformDotnetTestSupportToFalse>();
     }
 
     private static IServiceCollection PublishingSettings(this IServiceCollection services)